chore: moved components to base directory and renamed them
